Annotate V5 C compiler with comments.
All of my comments begin with "BCD:".  No other changes are intended other
than to document better how this version of the compiler works.

v5 was released in June, 1974.

This is the earliest known version of the original C compiler to use the
preprocessor, and is thus much easier to understand than its predecessors.
(v4 sources are not known to exist; the v4 manual does, but it does not
mention a preprocessor.)